BusinessCardPCB - Update commit - Hour 1
Getting started on learning PCB design and NFC inductance calculation. Attached in the progress file is the method used to calculate the inductance of the coil, and the capacitance of the tuning and the current looks on the pcb. The qr code silkscreen links to my GitHub profile. The gerber file attached is the PCB file and can be viewed on any gerber viewer